requirements.txt
python3-flask
python3-flask-httpauth
uwsgi
uwsgi-plugin-python3
